+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreStorage.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmorate.model.Genre;
 import java.util.Collection;
 
 public interface GenreStorage {
-    Genre getGenreById(Integer genreId);
+    Genre getGenreById(Long genreId);
 
     Collection<Genre> getAllGenres();
 }
